refactor: Move fragments to it's own subdirectory
