+++ app/src/main/java/com/openblocks/android/fragments/main/ModulesFragment.java
@@ -1,9 +1,8 @@
-package com.openblocks.android;
+package com.openblocks.android.fragments.main;
 
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 
@@ -11,6 +10,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.openblocks.android.R;
 import com.openblocks.android.adapters.ModulesRecyclerViewAdapter;
 import com.openblocks.android.modman.models.Module;
 import com.openblocks.moduleinterface.OpenBlocksModule;
