+++ app/src/main/java/com/openblocks/android/fragments/main/ProjectsFragment.java
@@ -1,4 +1,4 @@
-package com.openblocks.android;
+package com.openblocks.android.fragments.main;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.openblocks.android.R;
 
 /**
  * A simple {@link Fragment} subclass.
