+++ app/src/main/res/layout/fragment_projects.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProjectsFragment">
+    tools:context=".fragments.main.ProjectsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/projectRecyclerView"
